Fix testing for completeness.
